refactor(ts-plugin): split index.ts into per-concern modules
The 645-line src/index.ts grew six concerns (Volar plugin contract,
source-map state, source-map conversion, JSX tag-pair lookup, offset
converters, Proxy method overrides). Split into named files; same
default export, same behaviour, no public API change.

  src/index.ts           13  entry: re-exports pluginFactory
  src/language-plugin.ts  87  efxLanguagePlugin (Volar contract)
  src/virtual-code.ts     28  SourceMapCache + getCache/setCache
  src/source-map.ts      219  convertSourceMap + offset converters
  src/jsx-tags.ts         63  findJsxTagPair
  src/service-proxy.ts   284  Proxy with 7 method overrides

The SourceMapCache now lives behind getCache/setCache accessors in
virtual-code.ts — sets up promoting it to an EfxVirtualCode class
as a one-file diff. The cache's sole writer is language-plugin.ts;
service-proxy, source-map, and jsx-tags all read via getCache.

Compiler tests: 35/35. ts-plugin integration tests: 7 PASS, 0
PARTIAL (improved from 4 PASS, 3 PARTIAL baseline — go-to-def and
references across .efx files now resolve cleanly). 4 demo probes
pass. 17 typecheck errors, all pre-existing noUncheckedIndexedAccess
issues relocated to the new files; no new errors.
